+++ bevy_kot_ui/src/interaction/cursor_position.rs
@@ -32,7 +32,7 @@ impl<'w, 's, C: LunexCursor> CursorPos<'w, 's, C>
     pub fn get_world(&self) -> Option<Vec2>
     {
         let Ok(cursor) = self.cursor.get_single() else { return None; };
-        Some(*cursor.position_world())
+        Some(*cursor.location_world())
     }
 
     /// Get the cursor's screen position in the focused window (origin is top left).
@@ -41,7 +41,7 @@ impl<'w, 's, C: LunexCursor> CursorPos<'w, 's, C>
     pub fn get_screen(&self) -> Option<Vec2>
     {
         let Ok(cursor) = self.cursor.get_single() else { return None; };
-        Some(*cursor.position_screen())
+        Some(*cursor.location_screen())
     }
 
     //todo: request position for specific window