+++ examples/10k_buttons.rs
@@ -162,11 +162,11 @@ fn setup(mut commands: Commands, window: Query<Entity, (With<Window>, With<Prima
         );
 
     // make lunex cursor
-    commands.spawn((Cursor::new(0.0), Transform::default(), MainMouseCursor));
+    commands.spawn((Cursor::new(), Transform::default(), MainMouseCursor));
 
     // prepare lunex ui tree
-    commands.insert_resource(StyleStackRes::<MainUI>::default());
-    let tree = UiTree::<MainUI>::new("ui");
+    commands.insert_resource(StyleStackRes::<MainUi>::default());
+    let tree = UiTree::<MainUi>::new("ui");
 
     let window = window.single();
     commands.entity(window).insert(tree.bundle());
@@ -187,7 +187,7 @@ fn main()
             )
         )
         .add_plugins(FpsTrackerPlugin)
-        .add_plugins(LunexUiPlugin2D::<MainUI>::new())
+        .add_plugins(LunexUiPlugin2D::<MainUi>::new())
         //.add_plugins(UIDebugOverlayPlugin)
         .add_plugins(ReactPlugin)
         .insert_resource(WinitSettings{